+++ Bootstrap.php
@@ -26,7 +26,7 @@ class Shopware_Plugins_Frontend_jtlconnector_Bootstrap extends Shopware_Componen
 
     public function getVersion()
     {
-        return '1.1.1';
+        return '1.1.2';
     }
 
     public function getInfo()
@@ -171,6 +171,8 @@ class Shopware_Plugins_Frontend_jtlconnector_Bootstrap extends Shopware_Componen
             case '1.0.12':
                 break;
             case '1.1.0':
+                break;
+            case '1.1.1':
                 break;
             default:
                 return false;