+++ src/jtl/Connector/Shopware/Mapper/Category.php
@@ -6,6 +6,7 @@
 
 namespace jtl\Connector\Shopware\Mapper;
 
+use jtl\Connector\Core\Logger\Logger;
 use jtl\Connector\Shopware\Model\CategoryAttr;
 use \jtl\Connector\Shopware\Utilities\Mmc;
 use \jtl\Connector\Model\Category as CategoryModel;
@@ -28,14 +29,20 @@ class Category extends DataMapper
         return $this->Manager()->find('Shopware\Models\Category\Category', $id);
     }
 
-    public function findByNameAndLevel($name, $level = 0)
+    public function findByNameAndLevel($name, $parentId = null)
     {
+        $sql = ' AND c.parent IS NULL';
+        $params = array($name);
+        if ($parentId !== null) {
+            $sql = ' AND c.parent = ?';
+            $params[] = $parentId;
+        }
+
         $id = Shopware()->Db()->fetchOne(
             'SELECT c.id
               FROM s_categories c
-              JOIN jtl_connector_category_level l ON l.category_id = c.id
-              WHERE c.description = ? AND l.level = ?',
-            array($name, $level)
+              WHERE c.description = ?' . $sql,
+            $params
         );
 
         if ($id !== null && (int) $id > 0) {
@@ -206,7 +213,6 @@ class Category extends DataMapper
     public function save(CategoryModel $category)
     {
         $categorySW = null;
-        //$result = new CategoryModel;
         $result = $category;
 
         if ($category->getParentCategoryId() !== null && isset(self::$parentCategoryIds[$category->getParentCategoryId()->getHost()])) {
@@ -228,7 +234,7 @@ class Category extends DataMapper
             $this->prepareCategoryMapping($category, $categorySW);
         }
 
-        if ($categorySW !== null) {
+        if ($categorySW !== null && $categorySW->getId() > 0) {
             self::$parentCategoryIds[$category->getId()->getHost()] = $categorySW->getId();
         }
 
@@ -294,7 +300,7 @@ class Category extends DataMapper
             }
 
             if ($name !== null) {
-                $categorySW = $this->findByNameAndLevel($name, $category->getLevel());
+                $categorySW = $this->findByNameAndLevel($name, $parentId);
             }
         }