+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+1.1.2
+-----
+- Fixed product variation language bug
+- Fixed measurement language bug
+- Fixed category loading by name
+- Added product variation type (standard, swatches) support
+- Added product child image relation support
+
 1.1.1
 -----
 - Added category by name mapping